[data] validate http response
